requirements.txt
streamlit
pandas
matplotlib
numpy
requests
seaborn
Pillow
pickle5
lime
joblib
scikit-learn
plotly
streamlit-pandas-profiling
pyodbc
pymssql